+++ CHANGELOG.md
@@ -1,1 +1,13 @@
 ## HEAD
+
+## 0.2.0
+
+###### _Mar 23, 2017_
+
+###### Component Fixes / Enhancements
+
+- Corrected `fixed` width tabs so that no scrolling behavior occurs. Previously, selecting a partially out of view tab would scroll that tab into view. @shawnmcknight
+
+###### Examples
+
+- Created `/examples` directory with a webpack-dev-server based site that will show interactive examples and code for various tab permutations. Docs for use are [here](https://github.com/STORIS/material-ui-scrollable-tabs/blob/master/examples/README.md). @shawnmcknight